requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==4.11.0
django-configurations==0.8
django-debug-toolbar==1.3.0
django-dotenv==1.3.0
django-extensions==1.5.5
django-grappelli==2.6.2
django-secure==1.0.1
django-toolbelt==0.0.1
Django==1.7
gunicorn==19.3.0
ipython==2.3.1
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
mock==1.0.1
oauthlib==0.7.2
psycopg2==2.6
python-slugify==0.0.9
requests-oauthlib==0.4.2
requests==2.4.2
responses==0.3.0
six==1.9.0
static3==0.6.0
Unidecode==0.04.17
Werkzeug==0.10.4
whitenoise==1.0.5
